Rename initramfs package to vm
It is actually about running a cpud vm.
