+++ vm/build.go
